pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-model-security-analyzer"
version = "1.0.0"
description = "Static security analyzer for machine learning model files"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Security Team", email = "security@example.com"}
]
keywords = ["security", "machine-learning", "vulnerability-scanner", "ml-security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "huggingface_hub>=0.20.0",
    "protobuf>=4.21.0",
    "h5py>=3.8.0",
    "numpy>=1.24.0",
    "jsonschema>=4.17.0",
    "fickling>=0.1.0",
]

[project.optional-dependencies]
cloud = ["boto3>=1.28.0", "google-cloud-storage>=2.10.0"]
secrets = ["trufflehog3>=3.0.0"]
modelscan = ["modelscan>=0.8.0"]
guarddog = ["guarddog>=2.0.0"]
tensorflow = ["tensorflow>=2.13.0"]
dev = ["pytest>=7.3.0", "pytest-cov>=4.1.0", "black>=23.0.0", "mypy>=1.4.0"]
all = [
    "boto3>=1.28.0",
    "google-cloud-storage>=2.10.0",
    "trufflehog3>=3.0.0",
    "modelscan>=0.8.0",
    "guarddog>=2.0.0",
]

[project.scripts]
mlsec-analyzer = "mlsec_analyzer.cli:main"

[project.urls]
Homepage = "https://github.com/example/ml-model-security-analyzer"
Documentation = "https://github.com/example/ml-model-security-analyzer#readme"
Repository = "https://github.com/example/ml-model-security-analyzer"
Issues = "https://github.com/example/ml-model-security-analyzer/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
